Fix test suite to achieve 100% green status
- Fixed SetCommandHandler constructor to accept command_service parameter
- Updated streaming test to expect text/event-stream content-type
- Fixed ruff linting issues across source and test files
- Added force_reload parameter to OpenAI OAuth test mock
- Fixed BackendError import paths in multiple files
- Applied black formatting to all source and test files
- Updated help command snapshots with actual output
- Fixed pytest compression threshold test logic
- Resolved import organization and type annotation issues
- Fixed architectural violation in transport adapter test

+++ src/core/adapters/exception_adapters.py
@@ -35,7 +35,7 @@ def _build_retry_after_header(reset_at: float | None) -> dict[str, str] | None:
     if delay_seconds <= 0:
         return {"Retry-After": "0"}
 
-    return {"Retry-After": str(int(math.ceil(delay_seconds)))}
+    return {"Retry-After": str(math.ceil(delay_seconds))}
 
 
 def create_exception_handler() -> (
