+++ src/core/transport/fastapi/exception_adapters.py
@@ -40,7 +40,7 @@ def _build_retry_after_header(reset_at: float | None) -> dict[str, str] | None:
     if delay_seconds <= 0:
         return {"Retry-After": "0"}
 
-    return {"Retry-After": str(int(math.ceil(delay_seconds)))}
+    return {"Retry-After": str(math.ceil(delay_seconds))}
 
 
 def map_domain_exception_to_http_exception(exc: LLMProxyError) -> HTTPException:
